+++ iqm_data_definitions/src/iqm/data_definitions/common/v1/setting_pb2.py
@@ -15,7 +15,7 @@ from iqm.data_definitions.common.v1 import parameter_pb2 as iqm_dot_data__defini
 from iqm.data_definitions.common.v1 import data_types_pb2 as iqm_dot_data__definitions_dot_common_dot_v1_dot_data__types__pb2
 
 
-DESCRIPTOR = _descriptor_pool.Default().AddSerializedFile(b'\n,iqm/data_definitions/common/v1/setting.proto\x12\x1eiqm.data_definitions.common.v1\x1a.iqm/data_definitions/common/v1/parameter.proto\x1a/iqm/data_definitions/common/v1/data_types.proto\"\xcd\x04\n\x0bSettingNode\x12\x0c\n\x04name\x18\x01 \x01(\t\x12K\n\x08settings\x18\x02 \x03(\x0b\x32\x39.iqm.data_definitions.common.v1.SettingNode.SettingsEntry\x12K\n\x08subnodes\x18\x03 \x03(\x0b\x32\x39.iqm.data_definitions.common.v1.SettingNode.SubnodesEntry\x1a\xd1\x01\n\x07Setting\x12\x18\n\x0eparameter_name\x18\x01 \x01(\tH\x00\x12>\n\tparameter\x18\x02 \x01(\x0b\x32).iqm.data_definitions.common.v1.ParameterH\x00\x12\x34\n\x05value\x18\x03 \x01(\x0b\x32%.iqm.data_definitions.common.v1.Datum\x12\x16\n\tread_only\x18\x04 \x01(\x08H\x01\x88\x01\x01\x42\x10\n\x0eparameter_descB\x0c\n\n_read_only\x1a\x64\n\rSettingsEntry\x12\x0b\n\x03key\x18\x01 \x01(\t\x12\x42\n\x05value\x18\x02 \x01(\x0b\x32\x33.iqm.data_definitions.common.v1.SettingNode.Setting:\x02\x38\x01\x1a\\\n\rSubnodesEntry\x12\x0b\n\x03key\x18\x01 \x01(\t\x12:\n\x05value\x18\x02 \x01(\x0b\x32+.iqm.data_definitions.common.v1.SettingNode:\x02\x38\x01\x62\x06proto3')
+DESCRIPTOR = _descriptor_pool.Default().AddSerializedFile(b'\n,iqm/data_definitions/common/v1/setting.proto\x12\x1eiqm.data_definitions.common.v1\x1a.iqm/data_definitions/common/v1/parameter.proto\x1a/iqm/data_definitions/common/v1/data_types.proto\"\xa7\x04\n\x0bSettingNode\x12\x0c\n\x04name\x18\x01 \x01(\t\x12K\n\x08settings\x18\x02 \x03(\x0b\x32\x39.iqm.data_definitions.common.v1.SettingNode.SettingsEntry\x12K\n\x08subnodes\x18\x03 \x03(\x0b\x32\x39.iqm.data_definitions.common.v1.SettingNode.SubnodesEntry\x1a\xab\x01\n\x07Setting\x12\x18\n\x0eparameter_name\x18\x01 \x01(\tH\x00\x12>\n\tparameter\x18\x02 \x01(\x0b\x32).iqm.data_definitions.common.v1.ParameterH\x00\x12\x34\n\x05value\x18\x03 \x01(\x0b\x32%.iqm.data_definitions.common.v1.DatumB\x10\n\x0eparameter_desc\x1a\x64\n\rSettingsEntry\x12\x0b\n\x03key\x18\x01 \x01(\t\x12\x42\n\x05value\x18\x02 \x01(\x0b\x32\x33.iqm.data_definitions.common.v1.SettingNode.Setting:\x02\x38\x01\x1a\\\n\rSubnodesEntry\x12\x0b\n\x03key\x18\x01 \x01(\t\x12:\n\x05value\x18\x02 \x01(\x0b\x32+.iqm.data_definitions.common.v1.SettingNode:\x02\x38\x01\x62\x06proto3')
 
 _builder.BuildMessageAndEnumDescriptors(DESCRIPTOR, globals())
 _builder.BuildTopDescriptorsAndMessages(DESCRIPTOR, 'iqm.data_definitions.common.v1.setting_pb2', globals())
@@ -27,11 +27,11 @@ if _descriptor._USE_C_DESCRIPTORS == False:
   _SETTINGNODE_SUBNODESENTRY._options = None
   _SETTINGNODE_SUBNODESENTRY._serialized_options = b'8\001'
   _SETTINGNODE._serialized_start=178
-  _SETTINGNODE._serialized_end=767
+  _SETTINGNODE._serialized_end=729
   _SETTINGNODE_SETTING._serialized_start=362
-  _SETTINGNODE_SETTING._serialized_end=571
-  _SETTINGNODE_SETTINGSENTRY._serialized_start=573
-  _SETTINGNODE_SETTINGSENTRY._serialized_end=673
-  _SETTINGNODE_SUBNODESENTRY._serialized_start=675
-  _SETTINGNODE_SUBNODESENTRY._serialized_end=767
+  _SETTINGNODE_SETTING._serialized_end=533
+  _SETTINGNODE_SETTINGSENTRY._serialized_start=535
+  _SETTINGNODE_SETTINGSENTRY._serialized_end=635
+  _SETTINGNODE_SUBNODESENTRY._serialized_start=637
+  _SETTINGNODE_SUBNODESENTRY._serialized_end=729
 # @@protoc_insertion_point(module_scope)